refactor(landing): broke down landing page for maintainability
